Refactor into one common rule: `component-name-unique`
