+++ packages/core/src/rules/common/__tests__/component-name-unique.test.ts
@@ -1,0 +1,700 @@
+import { outdent } from 'outdent';
+import { parseYamlToDocument, replaceSourceWithRef } from '../../../../__tests__/utils';
+import { lintDocumentForTest } from './utils/lint-document-for-test';
+
+describe('Oas3 component-name-unique', () => {
+  describe('schema', () => {
+    it('should report on multiple schemas with same name', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+          openapi: 3.0.0
+          components:
+            schemas:
+              SomeSchema:
+                type: object
+              Test:
+                type: object
+                properties:
+                  there:
+                    $ref: '/test.yaml#/components/schemas/SomeSchema'
+        `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/test.yaml',
+          body: outdent`
+          openapi: 3.0.0
+          components:
+            schemas:
+              SomeSchema:
+                type: object
+        `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "location": Array [
+              Object {
+                "pointer": "#/",
+                "reportOnKey": false,
+                "source": "/foobar.yaml",
+              },
+            ],
+            "message": "Schema 'SomeSchema' is not unique. It is defined at:
+        - /foobar.yaml#/components/schemas/SomeSchema
+        - /test.yaml#/components/schemas/SomeSchema",
+            "ruleId": "component-name-unique",
+            "severity": "error",
+            "suggest": Array [],
+          },
+        ]
+      `);
+    });
+
+    it('should report on multiple schemas with same name - filename', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+        openapi: 3.0.0
+        components:
+          schemas:
+            SomeSchema:
+              type: object
+            Test:
+              type: object
+              properties:
+                there:
+                  $ref: '/SomeSchema.yaml'
+      `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/SomeSchema.yaml',
+          body: outdent`
+          type: object
+          properties:
+            test:
+              type: number
+      `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`
+      Array [
+        Object {
+          "location": Array [
+            Object {
+              "pointer": "#/",
+              "reportOnKey": false,
+              "source": "/foobar.yaml",
+            },
+          ],
+          "message": "Schema 'SomeSchema' is not unique. It is defined at:
+      - /foobar.yaml#/components/schemas/SomeSchema
+      - /SomeSchema.yaml",
+          "ruleId": "component-name-unique",
+          "severity": "error",
+          "suggest": Array [],
+        },
+      ]
+    `);
+    });
+
+    it('should not report on multiple schemas with different names', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+          openapi: 3.0.0
+          components:
+            schemas:
+              SomeSchema:
+                type: object
+              Test:
+                type: object
+                properties:
+                  there:
+                    $ref: '/test.yaml#/components/schemas/OtherSchema'
+        `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/test.yaml',
+          body: outdent`
+          openapi: 3.0.0
+          components:
+            schemas:
+              OtherSchema:
+                type: object
+        `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot('Array []');
+    });
+  });
+
+  describe('parameter', () => {
+    it('should report if multiple parameters have same component name', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+        openapi: 3.0.0
+        paths:
+          /test:
+            get:
+              parameters:
+              - $ref: '#/components/parameters/ParameterOne'
+              - $ref: '/test.yaml#/components/parameters/ParameterOne'
+        components:
+          parameters:
+            ParameterOne:
+              name: parameterOne
+              in: query
+              schema:
+                type: integer
+      `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/test.yaml',
+          body: outdent`
+        openapi: 3.0.0
+        components:
+          parameters:
+            ParameterOne:
+              name: oneParameter
+              in: query
+              schema:
+                type: integer
+        `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`
+      Array [
+        Object {
+          "location": Array [
+            Object {
+              "pointer": "#/",
+              "reportOnKey": false,
+              "source": "/foobar.yaml",
+            },
+          ],
+          "message": "Parameter 'ParameterOne' is not unique. It is defined at:
+      - /foobar.yaml#/components/parameters/ParameterOne
+      - /test.yaml#/components/parameters/ParameterOne",
+          "ruleId": "component-name-unique",
+          "severity": "error",
+          "suggest": Array [],
+        },
+      ]
+    `);
+    });
+
+    it('should report on multiple parameters with same component name - filename', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+        openapi: 3.0.0
+        paths:
+          /test:
+            get:
+              parameters:
+              - $ref: '#/components/parameters/ParameterOne'
+              - $ref: '/ParameterOne.yaml'
+        components:
+          parameters:
+            ParameterOne:
+              name: parameterOne
+              in: query
+              schema:
+                type: integer
+      `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/ParameterOne.yaml',
+          body: outdent`
+          name: oneParameter
+          in: query
+          schema:
+            type: integer
+        `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`
+      Array [
+        Object {
+          "location": Array [
+            Object {
+              "pointer": "#/",
+              "reportOnKey": false,
+              "source": "/foobar.yaml",
+            },
+          ],
+          "message": "Parameter 'ParameterOne' is not unique. It is defined at:
+      - /foobar.yaml#/components/parameters/ParameterOne
+      - /ParameterOne.yaml",
+          "ruleId": "component-name-unique",
+          "severity": "error",
+          "suggest": Array [],
+        },
+      ]
+    `);
+    });
+
+    it('should not report on multiple parameters with different component names', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+        openapi: 3.0.0
+        paths:
+          /test:
+            get:
+              parameters:
+              - $ref: '#/components/parameters/ParameterOne'
+              - $ref: '/test.yaml#/components/parameters/OneParameter'
+        components:
+          parameters:
+            ParameterOne:
+              name: parameterOne
+              in: query
+              schema:
+                type: integer
+      `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/test.yaml',
+          body: outdent`
+        openapi: 3.0.0
+        components:
+          parameters:
+            OneParameter:
+              name: oneParameter
+              in: query
+              schema:
+                type: integer
+      `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot('Array []');
+    });
+  });
+
+  describe('response', () => {
+    it('should report if multiple responses have same component name', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+        openapi: 3.0.0
+        paths:
+          /test:
+            get:
+              responses:
+                '200':
+                  $ref: '#/components/responses/SuccessResponse'
+          /test2:
+            get:
+              responses:
+                '200':
+                  $ref: '/test.yaml#/components/responses/SuccessResponse'
+        components:
+          responses:
+            SuccessResponse:
+              description: Successful response
+              content:
+                application/json:
+                  schema:
+                    type: string
+      `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/test.yaml',
+          body: outdent`
+        openapi: 3.0.0
+        components:
+          responses:
+            SuccessResponse:
+              description: Successful response
+              content:
+                application/json:
+                  schema:
+                    type: string
+        `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`
+      Array [
+        Object {
+          "location": Array [
+            Object {
+              "pointer": "#/",
+              "reportOnKey": false,
+              "source": "/foobar.yaml",
+            },
+          ],
+          "message": "Response 'SuccessResponse' is not unique. It is defined at:
+      - /foobar.yaml#/components/responses/SuccessResponse
+      - /test.yaml#/components/responses/SuccessResponse",
+          "ruleId": "component-name-unique",
+          "severity": "error",
+          "suggest": Array [],
+        },
+      ]
+    `);
+    });
+
+    it('should report on multiple responses with same component name - filename', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+        openapi: 3.0.0
+        paths:
+          /test:
+            get:
+              responses:
+                '200':
+                  $ref: '#/components/responses/SuccessResponse'
+          /test2:
+            get:
+              responses:
+                '200':
+                  $ref: '/SuccessResponse.yaml'
+        components:
+          responses:
+            SuccessResponse:
+              description: Successful response
+              content:
+                application/json:
+                  schema:
+                    type: string
+      `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/SuccessResponse.yaml',
+          body: outdent`
+          description: Successful response
+          content:
+            application/json:
+              schema:
+                type: string
+        `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`
+      Array [
+        Object {
+          "location": Array [
+            Object {
+              "pointer": "#/",
+              "reportOnKey": false,
+              "source": "/foobar.yaml",
+            },
+          ],
+          "message": "Response 'SuccessResponse' is not unique. It is defined at:
+      - /foobar.yaml#/components/responses/SuccessResponse
+      - /SuccessResponse.yaml",
+          "ruleId": "component-name-unique",
+          "severity": "error",
+          "suggest": Array [],
+        },
+      ]
+    `);
+    });
+
+    it('should not report on multiple responses with different component names', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+        openapi: 3.0.0
+        paths:
+          /test:
+            get:
+              responses:
+                '200':
+                  $ref: '#/components/responses/TestSuccessResponse'
+          /test2:
+            get:
+              responses:
+                '200':
+                  $ref: '/test.yaml#/components/responses/Test2SuccessResponse'
+        components:
+          responses:
+            TestSuccessResponse:
+              description: Successful response
+              content:
+                application/json:
+                  schema:
+                    type: string
+      `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/test.yaml',
+          body: outdent`
+        openapi: 3.0.0
+        components:
+          responses:
+            Test2SuccessResponse:
+              description: Successful response
+              content:
+                application/json:
+                  schema:
+                    type: string
+      `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot('Array []');
+    });
+  });
+
+  describe('request-body', () => {
+    it('should report if multiple request bodies have same component name', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+        openapi: 3.0.0
+        paths:
+          /test:
+            post:
+              requestBody:
+                $ref: '#/components/requestBodies/MyRequestBody'
+          /test2:
+            post:
+              requestBody:
+                $ref: '/test.yaml#/components/requestBodies/MyRequestBody'
+        components:
+          requestBodies:
+            MyRequestBody:
+              required: true
+              content:
+                application/json:
+                  schema:
+                    type: string
+      `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/test.yaml',
+          body: outdent`
+        components:
+          requestBodies:
+            MyRequestBody:
+              required: true
+              content:
+                application/json:
+                  schema:
+                    type: string
+        `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`
+      Array [
+        Object {
+          "location": Array [
+            Object {
+              "pointer": "#/",
+              "reportOnKey": false,
+              "source": "/foobar.yaml",
+            },
+          ],
+          "message": "RequestBody 'MyRequestBody' is not unique. It is defined at:
+      - /foobar.yaml#/components/requestBodies/MyRequestBody
+      - /test.yaml#/components/requestBodies/MyRequestBody",
+          "ruleId": "component-name-unique",
+          "severity": "error",
+          "suggest": Array [],
+        },
+      ]
+    `);
+    });
+
+    it('should report on multiple responses with same component name - filename', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+        openapi: 3.0.0
+        paths:
+          /test:
+            post:
+              requestBody:
+                $ref: '#/components/requestBodies/MyRequestBody'
+          /test2:
+            post:
+              requestBody:
+                $ref: '/MyRequestBody.yaml'
+        components:
+          requestBodies:
+            MyRequestBody:
+              required: true
+              content:
+                application/json:
+                  schema:
+                    type: string
+      `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/MyRequestBody.yaml',
+          body: outdent`
+          components:
+            requestBodies:
+              MyRequestBody:
+                required: true
+                content:
+                  application/json:
+                    schema:
+                      type: string
+        `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`
+      Array [
+        Object {
+          "location": Array [
+            Object {
+              "pointer": "#/",
+              "reportOnKey": false,
+              "source": "/foobar.yaml",
+            },
+          ],
+          "message": "RequestBody 'MyRequestBody' is not unique. It is defined at:
+      - /foobar.yaml#/components/requestBodies/MyRequestBody
+      - /MyRequestBody.yaml",
+          "ruleId": "component-name-unique",
+          "severity": "error",
+          "suggest": Array [],
+        },
+      ]
+    `);
+    });
+
+    it('should not report on multiple responses with different component names', async () => {
+      const document = parseYamlToDocument(
+        outdent`
+        openapi: 3.0.0
+        paths:
+          /test:
+            post:
+              requestBody:
+                $ref: '#/components/requestBodies/TestRequestBody'
+          /test2:
+            post:
+              requestBody:
+                $ref: '/test.yaml#/components/requestBodies/Test2RequestBody'
+        components:
+          requestBodies:
+            TestRequestBody:
+              required: true
+              content:
+                application/json:
+                  schema:
+                    type: string
+      `,
+        '/foobar.yaml'
+      );
+      const additionalDocuments = [
+        {
+          absoluteRef: '/test.yaml',
+          body: outdent`
+        components:
+          requestBodies:
+            Test2RequestBody:
+              required: true
+              content:
+                application/json:
+                  schema:
+                    type: string
+      `,
+        },
+      ];
+
+      const results = await lintDocumentForTest(
+        { 'component-name-unique': 'error' },
+        document,
+        additionalDocuments
+      );
+
+      expect(replaceSourceWithRef(results)).toMatchInlineSnapshot('Array []');
+    });
+  });
+});
